moved delete button to the edit component
