+++ react-app/src/components/Directions/ShowDirections.module.css
@@ -14,7 +14,7 @@
     margin-bottom: 10px;
 }
 
-.directions_step h4
+.directions_step h3
 {
     margin-bottom: 5px;
 }
